requirement.txt
redis
celery==4.0.1
python-dotenv
pymongo
requests
PyMySQL
